added tokenize and compare functions
